move two_class_predictions
