+++ syserol/classify.py
@@ -5,7 +5,8 @@ import pandas as pd
 import tensorly as tl
 from sklearn.model_selection import KFold
 from sklearn.model_selection import cross_val_predict
-from sklearn.metrics import r2_score, accuracy_score
+from sklearn.linear_model import LogisticRegressionCV
+from sklearn.metrics import r2_score, accuracy_score, confusion_matrix
 from sklearn.svm import SVC
 from tensorly.kruskal_tensor import kruskal_to_tensor
 from syserol.tensor import perform_CMTF
@@ -73,3 +74,55 @@ def SVM_2class_predictions(subjects_matrix):
     nv_accuracy = accuracy_score(Y, y_pred2)
 
     return cp_accuracy, nv_accuracy
+
+
+def two_way_classifications():
+    """ Predict classifications of subjects by progression (EC/VC vs TP/UP) or by viremia (EC/TP vs VC/UP) - Alter methods"""
+    # Import Luminex, Luminex-IGG, Subject group pairs, and Glycan into DF
+    df = importLuminex()
+    lum = df.pivot(index="subject", columns="variable", values="value")
+    subj = load_file("meta-subjects")
+    igg = importIGG()
+
+    igg = igg.pivot(index="subject", columns="variable", values="value")
+    data_frames = [lum, subj, igg]
+    df_merged = reduce(
+        lambda left, right: pd.merge(left, right, on=["subject"], how="inner"),
+        data_frames,
+    )
+    df_merged = df_merged.dropna()
+
+    # Subset, Z score
+    df_class = df_merged[["class.cp", "class.nv"]]
+    df_variables = df_merged.drop(
+        ["subject", "class.etuv", "class.cp", "class.nv"], axis=1
+    )
+    df_variables = df_variables.apply(zscore)
+
+    # Predict Controller vs. Progressor
+    Y1 = df_class["class.cp"]
+    Y1 = (Y1 == "controller").astype(int)  # controllers are 1s, progressors are 0s
+    X1 = df_variables
+
+    Y_pred1 = cross_val_predict(LogisticRegressionCV(), X1, Y1)
+    model1 = LogisticRegressionCV().fit(X1, Y1)
+
+    print(model1.coef_)
+    confusionCvP = confusion_matrix(Y1, Y_pred1)
+    accuracyCvP = accuracy_score(Y1, Y_pred1)
+    print(f"Confusion Matrix Controller vs. Progressor: {confusionCvP} \n")
+
+    # Predict Viremic vs. Nonviremic
+    Y2 = df_class["class.nv"]
+    Y2 = (Y2 == "viremic").astype(int)  # viremic = 1, nonviremic = 0
+    X2 = df_variables
+
+    Y_pred2 = cross_val_predict(LogisticRegressionCV(), X2, Y2)
+    model2 = LogisticRegressionCV().fit(X2, Y2)
+
+    print(model2.coef_)
+    confusionVvN = confusion_matrix(Y2, Y_pred2)
+    accuracyVvN = accuracy_score(Y2, Y_pred2)
+    print(f"Confusion Matrix Viremic vs. Nonviremic: {confusionVvN} \n")
+
+    return accuracyCvP, accuracyVvN, confusionCvP, confusionVvN
